Extract webhook Handlers into their own module
This moves the Webhook Handlers out of the WebhooksController; providing several benefits:

1. Webhook Handlers now live in their own module - they're not tied to the Webhooks Controller anymore.

2. The Webooks registry is out of the Rails autoloader path. In development it was quite often the case the the Rails reloader would re-initialize the extra_handlers attribute of the WebhooksController to an empty array - effectively un-registering all of the custom handlers. This avoids the case where a user would expect a custom handler to fire in development, but it has been effectively unregistered by the autoloader.

3. The interface between the defaut, shipit-engine supplied, and user defined handlers has been unified. A handler need only implement the callable interface.

4. Since handlers need only respond to call registering handlers becomes more flexible. Any PORO which implements a call method, however any other callable Ruby object will work just fine: e.g a block, a proc, a lambda, etc.

5. The registry allows for multiple handlers to be registered for the same event.
