+++ test/controllers/webhooks_controller_test.rb
@@ -154,16 +154,14 @@ module Shipit
     end
 
     test "other events trigger custom handlers" do
-      received_webhooks = []
-      Shipit::WebhooksController.register_handler do |type, params|
-        received_webhooks << [type, params]
-      end
+      event = 'pull_request'
+      mock_handler = mock
+      mock_handler.expects(:call).with(pull_request_params.stringify_keys).once
+      Shipit::Webhooks.register_handler(event, mock_handler)
 
-      @request.headers['X-Github-Event'] = 'pull_request'
+      @request.headers['X-Github-Event'] = event
       post :create, body: pull_request_params.to_json, as: :json
       assert_response :ok
-
-      assert_equal [['pull_request', pull_request_params.stringify_keys]], received_webhooks
     end
 
     private
